+++ src/pages/safelist/style.css
@@ -1,0 +1,132 @@
+html {
+	--transition: ease;
+	--fadetime: 0.15s;
+	--interact: #1da1f2;
+	--bg0color: #000000;
+	--bg1color: #1e1f25;
+	--bg2color: #151416;
+	--bg3color: var(--bordercolor);
+	--blockquote: var(--bordercolor);
+	--textcolor: #DDD;
+	--bordercolor: #2D333A;
+	--linecolor: var(--bordercolor);
+	--borderhover: var(--interact);
+	--subtle: #EEEEEE80;
+	--shadowcolor: #00000080;
+	--activeshadowcolor: #000000B3;
+	--screen-cover: #00000080;
+	--border-size: 1px;
+	--border-radius: 3px;
+
+	background: var(--bg0color);
+}
+html * {
+	font-family: Bitstream Vera Sans, DejaVu Sans, Arial, Helvetica, sans-serif;
+}
+body {
+	background: var(--bg1color);
+	color: var(--textcolor);
+}
+html, body, main, h1, p {
+	margin: 0;
+}
+
+::-webkit-scrollbar {
+	width: 12px;
+	height:12px;
+}
+::-webkit-scrollbar-track {
+	background: var(--bg2color);
+}
+::-webkit-scrollbar-thumb {
+	background: #4D535A;
+	border-radius: 4px;
+	border: 2px solid var(--bg2color);
+}
+::-webkit-scrollbar-thumb:hover {
+	background: var(--interact);
+}
+::-webkit-scrollbar-corner {
+	background: var(--bg0color);
+}
+
+main {
+	color: var(--textcolor);
+	text-align: center;
+	width: 100%;
+	background: #1E1F25;
+}
+.inner {
+	padding: 25px;
+	max-width: 30em;
+	margin: 0 auto;
+}
+.subtitle {
+	margin-bottom: 1em;
+}
+
+/* popup title */
+h2 {
+	margin: 1em 0 0;
+	text-align: center;
+}
+/* Blue Blocker logo in the popup title */
+h2 img {
+	height: 1.2em !important; /* overwrite inline style that was added for development */
+	margin-bottom: -0.15em;
+}
+
+.option {
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+}
+.option p {
+	text-align: center;
+	margin: 0;
+}
+
+input[type="file"] {
+	position: absolute;
+	display: none;
+	top: -100vh;
+	opacity: 0;
+	height: 0;
+	width: 0;
+}
+label {
+	cursor: pointer;
+	margin: 0;
+	display: inline-flex;
+	white-space: nowrap;
+	align-items: center;
+	padding-left: 0.29em;
+}
+label:hover,
+label:active,
+label:focus {
+	color: var(--interact);
+}
+
+.button,
+.button:active,
+.button:focus {
+	margin-left: 1em;
+	font-size: 1em;
+	padding: 0.5em 1em;
+	color: var(--textcolor);
+	background: var(--bg2color);
+	box-shadow: 0 2px 3px 1px var(--shadowcolor);
+	border: var(--border-size) solid var(--bordercolor);
+	border-radius: var(--border-radius);
+	-webkit-transition: var(--transition) var(--fadetime);
+	-moz-transition: var(--transition) var(--fadetime);
+	-o-transition: var(--transition) var(--fadetime);
+	transition: var(--transition) var(--fadetime);
+	cursor: pointer;
+}
+.button:hover {
+	box-shadow: 0 0 10px 3px var(--activeshadowcolor);
+	border-color: var(--interact);
+	color: var(--interact);
+}
